+++ CustomerData/CartTagging.php
@@ -85,7 +85,7 @@ class CartTagging extends HashedTagging implements SectionSourceInterface
     }
 
     /**
-     * @inheritdoc
+     * @inheritDoc
      */
     public function getSectionData()
     {