+++ Model/Person/Tagging/Builder.php
@@ -91,7 +91,7 @@ class Builder extends PersonBuilder
     }
 
     /**
-     * @inheritdoc
+     * @inheritDoc
      * @return Customer
      */
     public function buildObject(