+++ Model/ResourceModel/Product/Update/Queue.php
@@ -34,60 +34,23 @@
  *
  */
 
-namespace Nosto\Tagging\Model\Product\Cache;
+namespace Nosto\Tagging\Model\ResourceModel\Product\Update;
 
-use Magento\Catalog\Api\Data\ProductInterface;
-use Magento\Framework\Stdlib\DateTime\TimezoneInterface;
-use Magento\Store\Api\Data\StoreInterface;
-use Nosto\Tagging\Model\Product\Builder as NostoProductBuilder;
-use Nosto\Tagging\Model\Product\BuilderTrait;
-use Nosto\Tagging\Model\Product\Cache as CacheModel;
-use Nosto\Tagging\Model\Product\CacheFactory;
+use Magento\Framework\Model\ResourceModel\Db\AbstractDb;
+use Nosto\Tagging\Api\Data\ProductUpdateQueueInterface;
 
-class CacheBuilder
+class Queue extends AbstractDb
 {
-    use BuilderTrait {
-        BuilderTrait::__construct as builderTraitConstruct; // @codingStandardsIgnoreLine
-    }
+    protected $_serializableFields = [ProductUpdateQueueInterface::PRODUCT_IDS => [[], []]];
 
-    /** @var CacheFactory  */
-    private $cacheFactory;
-
-    /** @var NostoProductBuilder */
-    private $nostoProductBuilder;
-
-    /** @var TimezoneInterface */
-    private $magentoTimeZone;
-
+    const TABLE_NAME = 'nosto_tagging_product_update_queue';
     /**
-     * Builder constructor.
-     * @param CacheFactory $NostoCacheFactory
-     * @param TimezoneInterface $magentoTimeZone
+     * Initialize resource model
+     *
+     * @return void
      */
-    public function __construct(
-        CacheFactory $NostoCacheFactory,
-        TimezoneInterface $magentoTimeZone
-    ) {
-        $this->cacheFactory = $NostoCacheFactory;
-        $this->magentoTimeZone = $magentoTimeZone;
-    }
-
-    /**
-     * @param ProductInterface $product
-     * @param StoreInterface $store
-     * @return CacheModel
-     */
-    public function build(
-        ProductInterface $product,
-        StoreInterface $store
-    ) {
-        $productIndex = $this->cacheFactory->create();
-        $productIndex->setProductId($product->getId());
-        $productIndex->setCreatedAt($this->magentoTimeZone->date());
-        $productIndex->setInSync(false);
-        $productIndex->setIsDirty(true);
-        $productIndex->setUpdatedAt($this->magentoTimeZone->date());
-        $productIndex->setStore($store);
-        return $productIndex;
+    public function _construct()
+    {
+        $this->_init(self::TABLE_NAME, ProductUpdateQueueInterface::ID);
     }
 }