+++ Model/Service/Sync/Upsert/AsyncBulkPublisher.php
@@ -44,21 +44,33 @@ class AsyncBulkPublisher extends AbstractBulkPublisher
     const NOSTO_SYNC_MESSAGE_QUEUE = 'nosto_product_sync.update';
     const BULK_SIZE = 100;
 
+    /**
+     * @inheritDoc
+     */
     public function getTopicName(): string
     {
         return self::NOSTO_SYNC_MESSAGE_QUEUE;
     }
 
+    /**
+     * @inheritDoc
+     */
     public function getBulkSize(): int
     {
         return self::BULK_SIZE;
     }
 
+    /**
+     * @inheritDoc
+     */
     public function getBulkDescription(): string
     {
         return sprintf('Sync %d Nosto products', 2);
     }
 
+    /**
+     * @inheritDoc
+     */
     public function getMetaData(): string
     {
         return 'Sync Nosto products';