+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
 # not released
+* Refactor the indexing logic to use batched queues
+* Use Magento's built-in caching logic for caching Nosto product data 
 * Add google category as customisable attribute
 * Change the namespaces to comply with PHP SDK 5.0.0